requirements.txt
fastapi[standard]
pymongo
python-jose
passlib
pytest
python-dotenv
python-multipart
email-validator
bcrypt
motor